Move instrumentation operators to their own package.
